chore: refactor for increased modularity/future extensibility
extract cli-related functionality to cli dir

chore: refactor for increased modularity/extensibility
