+++ package.json
@@ -85,8 +85,7 @@
     "rollup-plugin-typescript2": "^0.21.1",
     "sade": "^1.4.2",
     "tiny-glob": "^0.2.6",
-    "ts-jest": "^24.0.2",
-    "tslib": "^1.9.3"
+    "ts-jest": "^24.0.2"
   },
   "devDependencies": {
     "@types/ansi-escapes": "^4.0.0",
@@ -100,6 +99,7 @@
     "@types/node": "^12.0.2",
     "@types/ora": "^3.2.0",
     "@types/react": "^16.8.17",
+    "@types/rollup": "^0.54.0",
     "@types/rollup-plugin-json": "^3.0.2",
     "@types/rollup-plugin-sourcemaps": "^0.4.2",
     "husky": "^2.3.0",
@@ -107,6 +107,7 @@
     "ps-tree": "^1.2.0",
     "react": "^16.8.6",
     "shelljs": "^0.8.3",
+    "tslib": "^1.10.0",
     "typescript": "^3.6.2"
   },
   "peerDependencies": {
