+++ src/builders/createRollupConfig.ts
@@ -1,8 +1,5 @@
-import { safeVariableName, safePackageName, external } from '../helpers';
-import { paths } from '../constants';
 import { terser } from 'rollup-plugin-terser';
 import { DEFAULT_EXTENSIONS } from '@babel/core';
-// import babel from 'rollup-plugin-babel';
 import commonjs from 'rollup-plugin-commonjs';
 import json from 'rollup-plugin-json';
 import replace from 'rollup-plugin-replace';
@@ -10,9 +7,10 @@ import resolve from 'rollup-plugin-node-resolve';
 import sourceMaps from 'rollup-plugin-sourcemaps';
 import typescript from 'rollup-plugin-typescript2';
 import { extractErrors } from '../errors/extractErrors';
-import { babelPluginTsdx } from '../babelPluginTsdx';
+import { babelPluginTsdx } from './createTsdxBabelPlugin';
 import { TsdxOptions } from '../types';
-
+import { safeVariableName, safePackageName, external } from '../helpers';
+import { paths } from '../constants';
 const errorCodeOpts = {
   errorMapFilePath: paths.appErrorsJson,
 };
