+++ src/builders/createTsdxConfig.ts
@@ -1,14 +1,15 @@
-import { TemplateStaticsBuilderOptions } from '../types';
+import { TsdxBag } from '../types';
+import { paths } from '../constants';
 import * as fs from 'fs-extra';
 
-export function getTsdxConfig(opts: TemplateStaticsBuilderOptions) {
+export function getTsdxConfig(opts: TsdxBag) {
   let tsdxConfig = {
     rollup(config: any, _options: any) {
       return config;
     },
   };
-  if (fs.existsSync(opts.paths.appConfig)) {
-    tsdxConfig = require(opts.paths.appConfig);
+  if (fs.existsSync(paths.appConfig)) {
+    tsdxConfig = require(paths.appConfig);
   }
   return tsdxConfig;
 }
