+++ src/cli/commands/index.ts
@@ -1,0 +1,5 @@
+export * from './build';
+export * from './create';
+export * from './test';
+export * from './watch';
+export * from './lint';
