+++ src/cli/utils/getDeps.ts
@@ -1,0 +1,13 @@
+import { commonDeps } from '../../constants';
+import { chromeDeps, reactDeps, basicDeps } from '../../generators';
+import { TemplateTypes } from '../../types';
+
+const depsStrategy = {
+  chrome: chromeDeps,
+  react: reactDeps,
+  basic: basicDeps,
+};
+
+export function getDeps(tempalte: TemplateTypes) {
+  return [...commonDeps, ...depsStrategy[tempalte]];
+}
