+++ src/cli/utils/getInstallArgs.ts
@@ -1,9 +1,6 @@
 import { InstallCommand } from './getInstallCmd';
 
-export default function getInstallArgs(
-  cmd: InstallCommand,
-  packages: string[]
-) {
+export function getInstallArgs(cmd: InstallCommand, packages: string[]) {
   switch (cmd) {
     case 'npm':
       return ['install', ...packages, '--save-dev'];
@@ -11,3 +8,5 @@ export default function getInstallArgs(
       return ['add', ...packages, '--dev'];
   }
 }
+
+export default getInstallArgs;
