+++ src/cli/utils/index.ts
@@ -1,0 +1,11 @@
+export * from './getInstallArgs';
+export * from './getInstallCmd';
+export * from './logError';
+export * from './normalizeOptions';
+export * from './fixGitignore';
+export * from './getSpinner';
+export * from './getProjectPath';
+export * from './copyTemplate';
+export * from './getTsdxBag';
+export * from './getTemplate';
+export * from './getDeps';
