+++ src/cli/utils/logError.ts
@@ -2,7 +2,7 @@ import chalk from 'chalk';
 
 const stderr = console.error.bind(console);
 
-export default function logError(err: any) {
+export function logError(err: any) {
   const error = err.error || err;
   const description = `${error.name ? error.name + ': ' : ''}${error.message ||
     error}`;
@@ -29,3 +29,5 @@ export default function logError(err: any) {
 
   stderr();
 }
+
+export default logError;
