+++ src/constants.ts
@@ -1,4 +1,11 @@
-import { resolveApp } from './helpers';
+import * as fs from 'fs-extra';
+import path from 'path';
+
+const appDirectory = fs.realpathSync(process.cwd());
+
+const resolveApp = function(relativePath: string) {
+  return path.resolve(appDirectory, relativePath);
+};
 
 export const paths = {
   appPackageJson: resolveApp('package.json'),
@@ -10,3 +17,13 @@ export const paths = {
   appDist: resolveApp('dist'),
   appConfig: resolveApp('tsdx.config.js'),
 };
+
+export const commonDeps = [
+  '@types/jest',
+  'husky',
+  'tsdx',
+  'tslib',
+  'typescript',
+].sort();
+
+export const defaultVersion = `0.1.0`;
